Update bootstrap.client.js
With this modification it is possible to use bootstrap with `$boostrap` this fixes some problem for exmaple with bootstrap scrollspy when changing from one page to another (in the old way and using data-bs-*) it will stop working. But with this you can apply scrollspy in mounted with javascript as the example says.